api/requirements.txt
amqp==2.6.1
asgiref==3.5.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==4.3.0
channels==4.0.0
Django==4.1.3
django-filter==22.1
djangorestframework==3.14.0
kombu==4.6.11
pytz==2022.6
sqlparse==0.4.3
vine==1.3.0
